pyproject.toml
[project]
name = "content-planner"
version = "0.1.0"
description = "Interactive AI-powered content idea generation tool for e-commerce products"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "ollama>=0.4.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
]

[project.scripts]
content-planner = "main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["agents", "ai", "ui", "utils", "main.py"]
